+++ android/app/src/main/java/com/now/naaga/common/dialog/NaagaAlertDialog.kt
@@ -1,4 +1,4 @@
-package com.now.naaga.presentation.onadventure
+package com.now.naaga.common.dialog
 
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
